Dialog: move files, add a story
