+++ src/Dialog/__tests__/Dialog.test.tsx
@@ -1,9 +1,9 @@
 import React, {useState, useRef} from 'react'
-import {Dialog, Box, Text} from '..'
-import {Button} from '../deprecated'
+import {Dialog, Box, Text} from '../..'
+import {Button} from '../../deprecated'
 import {render as HTMLRender, act, fireEvent} from '@testing-library/react'
 import {axe, toHaveNoViolations} from 'jest-axe'
-import {behavesAsComponent, checkExports} from '../utils/testing'
+import {behavesAsComponent, checkExports} from '../../utils/testing'
 expect.extend(toHaveNoViolations)
 
 const comp = (
