+++ src/Dialog/__tests__/Dialog.types.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Dialog from '../Dialog'
+import Dialog from '..'
 
 export function shouldAcceptCallWithNoProps() {
   return <Dialog />
