+++ src/Dialog/index.ts
@@ -1,0 +1,1 @@
+export {default, DialogProps} from './Dialog'
